Fixed inconsistencies in API Base, and revised PEP8/Pylint styles in API
